Add invalid swarm.best_pos check (#475)
This adds both a check whether best_pos is shapeless, as well as changing
the default initial value of best_cost so that no issues arise when copy()
is called on it without it having been modified.

+++ pyswarms/backend/swarms.py
@@ -100,7 +100,7 @@ class Swarm(object):
         validator=instance_of(np.ndarray),
     )
     best_cost = attrib(
-        type=float, default=np.inf, validator=instance_of((int, float))
+        type=float, default=np.float64(np.inf), validator=instance_of((int, float))
     )
     current_cost = attrib(
         type=np.ndarray,
